Extract shared refToUrl helper and handle remote refs in media entries
useModuleMediaEntries.getUrl ignored remote refs entirely, so richtext
gallery thumbnails for remote files pointed at the wrong URL. For draft
remote files the patch_id URL also needs the &remote=true&ref=... query
parameters that convertRemoteSource emits, otherwise the server cannot
fetch the binary from the remote bucket. Extract a shared refToUrl helper
so ModuleGallery and useModuleMediaEntries share the same logic.
